requirements.txt
openai
sounddevice
soundfile
numpy
python-dotenv
streamlit
pygame
pandas
plotly
